refactor(client, routes, country_list_item): literal errors corrected, changes in client, list_item
