+++ src/details/index.tsx
@@ -9,3 +9,6 @@ export const Details = () => {
         </div>
     );
 };
+
+
+
