+++ src/home/country_list_item/index.tsx
@@ -7,15 +7,12 @@ export const CountryListItem: FC<SharedTypes.ICountryOutput> = memo((item) => {
     const dispatch = useDispatch();
     const { deleteCountry, showCountryDetailes } = AppStore.Actions;
 
-    const onAddDetailes = (e: React.PointerEvent<HTMLButtonElement>): void => {
-        let data = e.currentTarget.dataset.lineId;
-        if (data) dispatch(showCountryDetailes(data));
-    };
-
-    const onDelete = (e: React.PointerEvent<HTMLButtonElement>): void => {
-        let data = e.currentTarget.dataset.lineId;
-        if (data) dispatch(deleteCountry(data));
-    };
+    const handleClick =
+        (action: (data: string) => AppStore.Actions.ActionsType /* any */) =>
+        (e: React.PointerEvent<HTMLButtonElement>): void => {
+            let data = e.currentTarget.dataset.lineId;
+            if (data) dispatch(action(data));
+        };
 
     const { flags, name, ccn3, region } = item;
 
@@ -28,16 +25,16 @@ export const CountryListItem: FC<SharedTypes.ICountryOutput> = memo((item) => {
             <SharedComponents.Text text={`${ccn3}`} />
             <SharedComponents.Text text={`${region}`} />
             <SharedComponents.LinkButton
-                ariaLabel="Detailes about country"
+                ariaLabel="Details about country"
                 text="...more"
-                to={Paths.DETAILES}
-                onClick={onAddDetailes}
+                to={Paths.DETAILS}
+                onClick={handleClick(showCountryDetailes)}
                 dataLineID={`${ccn3}`}
             />
             <SharedComponents.Button
                 ariaLabel="Remove country from list"
                 text="X"
-                onClick={onDelete}
+                onClick={handleClick(deleteCountry)}
                 dataLineID={`${ccn3}`}
             />
         </SharedComponents.HomepageTableLineContainer>
