+++ src/shared/store/client/index.tsx
@@ -1,18 +1,24 @@
 import { SharedTypes } from '@shared';
 
 export class _Client {
-    baseUrl?: string;
+    baseUrl: string;
+    fields?: Array<string>;
 
-    constructor({ baseUrl }: SharedTypes.IApiClientParams) {
+    constructor({ baseUrl, fields }: SharedTypes.IApiClientParams) {
+        if (baseUrl.trim() === '') {
+            throw new Error ('BaseURL is not defined');
+        }
         this.baseUrl = baseUrl;
+        this.fields = fields;
     }
 
     fetchCountryData = async ({
         code,
     }: SharedTypes.ICountryInput): Promise<SharedTypes.ICountryOutput[]> => {
         const countryData = await fetch(
-            `${this.baseUrl}/name/${code}?fields=area,borders,capital,capitalInfo,car,cca2,ccn3,currencies,flags,idd,languages,name,region,timezones,tld,coatOfArms`,
-            //берем только те поля, которые нам нужны: fields
+            this.fields
+                ? `${this.baseUrl}/name/${code}?fields=${this.fields.join(',')}`
+                : `${this.baseUrl}/name/${code}`,
         );
         return await countryData.json();
     };
@@ -20,4 +26,22 @@ export class _Client {
 
 export const Client = new _Client({
     baseUrl: 'https://restcountries.com/v3.1',
+    fields: [
+        'area',
+        'borders',
+        'capital',
+        'capitalInfo',
+        'car',
+        'cca2',
+        'ccn3',
+        'coatOfArms',
+        'currencies',
+        'flags',
+        'idd',
+        'languages',
+        'name',
+        'region',
+        'timezones',
+        'tld',
+    ],
 });
